Move dialogs out of src/components/Dialog/
Since the directory should only contain building blocks *for building* dialogs, not dialogs themselves.
