+++ src/components/TradeAsset/TradeAssetDialog.tsx
@@ -15,10 +15,10 @@ import * as routes from "../../routes"
 import ScrollableBalances from "../AccountAssets/ScrollableBalances"
 import { VerticalMargin } from "../Layout/Spacing"
 import MainTitle from "../MainTitle"
-import TradingForm from "../TradeAsset/TradingForm"
+import TradingForm from "./TradingForm"
 import TransactionSender from "../TransactionSender"
-import DialogBody from "./DialogBody"
-import { ActionButton, DialogActionsBox } from "./Generic"
+import DialogBody from "../Dialog/DialogBody"
+import { ActionButton, DialogActionsBox } from "../Dialog/Generic"
 import Portal from "../Portal"
 import { HorizontalLayout } from "../Layout/Box"
