+++ src/pages/all-accounts.tsx
@@ -3,12 +3,12 @@ import useMediaQuery from "@material-ui/core/useMediaQuery"
 import Button from "@material-ui/core/Button"
 import IconButton from "@material-ui/core/IconButton"
 import SettingsIcon from "@material-ui/icons/Settings"
+import AccountList from "../components/AccountList"
 import DialogBody from "../components/Dialog/DialogBody"
-import TermsAndConditions from "../components/Dialog/TermsAndConditions"
 import { Box } from "../components/Layout/Box"
 import { Section } from "../components/Layout/Page"
-import AccountList from "../components/AccountList"
 import MainTitle from "../components/MainTitle"
+import TermsAndConditions from "../components/TermsAndConditionsDialog"
 import { AccountsContext } from "../context/accounts"
 import { SettingsContext } from "../context/settings"
 import { useIsMobile, useRouter } from "../hooks/userinterface"
@@ -75,4 +75,4 @@ function AllAccountsPage() {
   )
 }
 
-export default AllAccountsPage
+export default React.memo(AllAccountsPage)
